pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Kurligin <kurligindevelopment@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
sqlalchemy = ">=2.0.41,<3.0.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = ">=2.11.7,<3.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
greenlet = ">=3.2.3,<4.0.0"
mcrcon = ">=0.7.0,<0.8.0"
python-dotenv = ">=1.1.1,<2.0.0"
javaproperties = ">=0.8.2,<0.9.0"
nbtlib = "^2.0.4"
aiodocker = "^0.24.0"
taskiq-fastapi = "^0.3.5"
taskiq = "^0.11.18"
taskiq-aio-pika = "^0.4.3"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["src"]
